requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
blas=1.0=mkl
ca-certificates=2024.8.30=hbcca054_0
cachetools=5.3.1=pypi_0
certifi=2023.7.22=pypi_0
charset-normalizer=3.2.0=pypi_0
click=7.1.2=pypi_0
cmake=3.27.5=pypi_0
comm=0.2.2=pyhd8ed1ab_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.12.1=pypi_0
datasets=2.18.0=pypi_0
debugpy=1.6.7=py38h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
diffusers=0.30.3=pypi_0
dill=0.3.8=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=2.1.0=pyhd8ed1ab_0
filelock=3.12.4=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.50.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2023.9.1=pypi_0
fuzzy-c-means=1.6.3=pypi_0
gast=0.4.0=pypi_0
giflib=5.2.1=h5eee18b_3
google-auth=2.23.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.58.0=pypi_0
h5py=3.9.0=pypi_0
huggingface-hub=0.25.0=pypi_0
idna=3.4=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.29.5=pyh3099207_0
ipython=8.12.0=pyh41d4057_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.2=pypi_0
jpeg=9e=h5eee18b_1
jsonlines=4.0.0=pypi_0
jupyter_client=7.3.4=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
keras=2.13.1=pypi_0
kiwisolver=1.4.5=pypi_0
kornia=0.6.4=pypi_0
latent-diffusion=0.0.1=dev_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libclang=16.0.6=pypi_0
libdeflate=1.17=h5eee18b_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libpng=1.6.39=h5eee18b_0
libretranslatepy=2.1.1=pypi_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libuv=1.44.2=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lightning-utilities=0.11.9=pypi_0
lit=16.0.6=pypi_0
loguru=0.7.2=pypi_0
lxml=5.1.0=pypi_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.4=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.20.3=pypi_0
numpy-base=1.19.2=py38h4c65ebe_1
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.4.127=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.3.0=pypi_0
openssl=1.1.1l=h7f98852_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=9.4.0=py38h6a678d5_0
pip=20.3.3=py38h06a4308_0
platformdirs=4.3.6=pyhd8ed1ab_0
prompt-toolkit=3.0.47=pyha770c72_0
prompt_toolkit=3.0.47=hd8ed1ab_0
protobuf=4.24.3=pypi_0
psutil=5.9.0=py38h5eee18b_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.3=pyhd8ed1ab_0
pyarrow=15.0.2=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydantic=1.10.14=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyparsing=3.1.2=pypi_0
python=3.8.5=h7579374_1
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.8=2_cp38
pytorch=1.7.0=py3.8_cuda11.0.221_cudnn8.0.3_0
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py38h6a678d5_0
readline=8.2=h5eee18b_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.5=pypi_0
scann=1.2.10=pypi_0
scikit-image=0.21.0=pypi_0
scipy=1.10.1=pypi_0
setuptools=68.0.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.8.10=pypi_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorflow=2.13.0=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=2.3.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.20.3=pypi_0
torch=2.3.0=pypi_0
torch-fidelity=0.3.0=pypi_0
torchmetrics=1.5.2=pypi_0
torchtext=0.18.0=pypi_0
torchvision=0.8.1=py38_cu110
tornado=6.1=py38h0a891b7_3
tqdm=4.66.1=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
transformers=4.46.3=pypi_0
translate=3.6.1=pypi_0
translate-toolkit=3.12.2=pypi_0
triton=2.3.0=pypi_0
typer=0.3.2=pypi_0
typing-extensions=4.11.0=pypi_0
tzdata=2024.1=pypi_0
urllib3=1.26.16=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=2.3.7=pypi_0
wget=3.2=pypi_0
wheel=0.38.4=py38h06a4308_0
wrapt=1.15.0=pypi_0
xxhash=3.4.1=pypi_0
xz=5.4.2=h5eee18b_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h6a678d5_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
